Update version_4.py, fix version_2.py and version_3.py
I have finished creating version_4.py. This version was an aesthetics enhancement version. I have also found a problem in versions 2 and 3 which was resolved by removing the while loop when adding or removing items from the basket. In version 4 there is now a message box that pops up and I changed what some of the buttons do.